requirements.txt
requests==2.32.3
PyYAML==6.0.2
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
colorama==0.4.6
tqdm==4.66.5